Make WindowingService not a singleton and improve GL helper lib
* Adds WindowingServiceData to expose global windowing backend
platform-specific data similar to how WindowData contains
platform-specific data for a native window.

* Adds InitializeOpenGl to WindowingService to allow OpenGL
initialization without creating a window that's OpenGL enabled.
Specifically this is needed to allow using OpenWindow.GL before
creating a window on Wayland because the EGL display connection
is required for some calls (e.g. SetSwapInterval).

* Require OpenWindowGl.Initialize call before usage (also because
of EGL display connection). I might change this later, because
it's not ideal to require an initialization call on a static class
before any other usage.